Extract terrain overlay mesh code resolver
